requirements.txt
django==3.1.4
django-rest-framework==0.1.0
django-filter==2.4.0